Add auto_reset_forces parameter for Simbody
The auto_reset_forces parameter allows the user to configure if the forces and torques applied to the bodies present in a simulation must be cleared out before the next step.

The default value for this parameter is true, in order to allow with the previous implementation.

Name and implementation of the parameter is based on DART physics implementation.